Enable and fix a bunch more lints

+++ egui_glow/src/painter.rs
@@ -684,10 +684,10 @@ pub fn clear(gl: &glow::Context, screen_size_in_pixels: [u32; 2], clear_color: e
         if true {
             // verified to be correct on eframe native (on Mac).
             gl.clear_color(
-                clear_color[0] as f32,
-                clear_color[1] as f32,
-                clear_color[2] as f32,
-                clear_color[3] as f32,
+                clear_color[0],
+                clear_color[1],
+                clear_color[2],
+                clear_color[3],
             );
         } else {
             let clear_color: Color32 = clear_color.into();
